Splitting GuardsService project to Exceptions and Interfaces. SourceFormatNodeRepository.Delete refactor, especially exception management
