+++ EncyclopediaGalactica.sln
@@ -108,6 +108,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "RestApi", "Host\RestApi\Res
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Controllers.ViewModels", "SourceFormats\Controllers.ViewModels\Controllers.ViewModels.csproj", "{5FE1733A-AE9F-494B-AD2E-C0DA7419C2F4}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GuardsService.Exceptions", "Utils\GuardsService.Exceptions\GuardsService.Exceptions.csproj", "{1079F6F8-0CF8-436B-817B-0B32A9318791}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GuardsService.Interfaces", "Utils\GuardsService.Interfaces\GuardsService.Interfaces.csproj", "{AE39BF10-5007-4E4E-AA87-34697E2C2842}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -298,6 +302,14 @@ Global
 		{5FE1733A-AE9F-494B-AD2E-C0DA7419C2F4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5FE1733A-AE9F-494B-AD2E-C0DA7419C2F4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5FE1733A-AE9F-494B-AD2E-C0DA7419C2F4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1079F6F8-0CF8-436B-817B-0B32A9318791}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1079F6F8-0CF8-436B-817B-0B32A9318791}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1079F6F8-0CF8-436B-817B-0B32A9318791}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1079F6F8-0CF8-436B-817B-0B32A9318791}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE39BF10-5007-4E4E-AA87-34697E2C2842}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE39BF10-5007-4E4E-AA87-34697E2C2842}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE39BF10-5007-4E4E-AA87-34697E2C2842}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE39BF10-5007-4E4E-AA87-34697E2C2842}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{E5435A2E-9B93-4F3C-80A6-49FCB4D1A1CB} = {37C84772-74FE-4A8A-8C5A-528040202EBD}
@@ -346,5 +358,7 @@ Global
 		{31D43021-281A-4C53-B308-03C92715C034} = {37C84772-74FE-4A8A-8C5A-528040202EBD}
 		{BAD10416-7A8C-4443-BF4E-C0E6E3F65916} = {93832915-FE03-4814-AE17-A9EF5848A82F}
 		{5FE1733A-AE9F-494B-AD2E-C0DA7419C2F4} = {37C84772-74FE-4A8A-8C5A-528040202EBD}
+		{1079F6F8-0CF8-436B-817B-0B32A9318791} = {A68E61C6-5EE5-4D53-ADCA-51B667BFA9F3}
+		{AE39BF10-5007-4E4E-AA87-34697E2C2842} = {A68E61C6-5EE5-4D53-ADCA-51B667BFA9F3}
 	EndGlobalSection
 EndGlobal
