+++ Host/RestApi/Program.cs
@@ -16,6 +16,7 @@ using EncyclopediaGalactica.SourceFormats.SourceFormatsService.Interfaces.Source
 using EncyclopediaGalactica.SourceFormats.SourceFormatsService.SourceFormatNodeService;
 using EncyclopediaGalactica.SourceFormats.ValidatorService;
 using EncyclopediaGalactica.Utils.GuardsService;
+using EncyclopediaGalactica.Utils.GuardsService.Interfaces;
 using FluentValidation.AspNetCore;
 using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
