+++ SourceFormats/E2E/SourceFormatNode/SourceFormatNodeSdk_Should.cs
@@ -1,4 +1,4 @@
-namespace EncyclopediaGalactica.SourceFormats.E2E;
+namespace EncyclopediaGalactica.SourceFormats.E2E.SourceFormatNode;
 
 using System.Diagnostics.CodeAnalysis;
 using Xunit;
