+++ SourceFormats/E2E/SourceFormatWebApplicationFactory.cs
@@ -27,6 +27,7 @@ using SourceFormatsService.Interfaces;
 using SourceFormatsService.Interfaces.SourceFormatNode;
 using SourceFormatsService.SourceFormatNodeService;
 using Utils.GuardsService;
+using Utils.GuardsService.Interfaces;
 using ValidatorService;
 
 [ExcludeFromCodeCoverage]
