+++ SourceFormats/SourceFormatsRepository.Int.Tests/SourceFormatNode/AddChildNodeValidationShould.cs
@@ -4,7 +4,7 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 using FluentAssertions;
-using Utils.GuardsService;
+using Utils.GuardsService.Exceptions;
 using Xunit;
 
 [ExcludeFromCodeCoverage]
@@ -26,7 +26,7 @@ public class AddChildNodeValidationShould : BaseTest
 
         // Assert
         await action.Should()
-            .ThrowExactlyAsync<GuardsServiceException>()
+            .ThrowExactlyAsync<GuardsServiceValueShouldNotBeEqualToException>()
             .ConfigureAwait(false);
     }
 }
