+++ SourceFormats/SourceFormatsRepository.Interfaces/ISourceFormatNodeRepository.cs
@@ -4,6 +4,7 @@ using Entities;
 using Exceptions;
 using FluentValidation;
 using Microsoft.EntityFrameworkCore;
+using Utils.GuardsService.Exceptions;
 
 /// <summary>
 ///     SourceFormatRepository interface for managing stored data in the database.
@@ -85,9 +86,14 @@ public interface ISourceFormatNodeRepository
     /// <param name="cancellationToken">
     ///     <see cref="CancellationToken" />
     /// </param>
-    /// <exception cref="SourceFormatNodeRepositoryException">
-    ///     In case any error and its enclosed exceptions provide
-    ///     additional information about the error.
+    /// <exception cref="InvalidOperationException">
+    ///     When no such entity
+    /// </exception>
+    /// <exception cref="ArgumentNullException">
+    ///     When null input provided.
+    /// </exception>
+    /// <exception cref="GuardsServiceValueShouldNotBeEqualToException">
+    ///     When the provided Id value is 0
     /// </exception>
     /// <returns>
     ///     Returns <see cref="Task" /> representing asynchronous operation.
@@ -103,6 +109,15 @@ public interface ISourceFormatNodeRepository
     /// <exception cref="SourceFormatNodeRepositoryException">
     ///     In case of any error and its inner exceptions provide additional details
     ///     about the error.
+    /// </exception>
+    /// <exception cref="OperationCanceledException">
+    ///     When operation is cancelled by a <see cref="CancellationToken" />
+    /// </exception>
+    /// <exception cref="DbUpdateException">
+    ///     When data cannot be persisted into database for some reason.
+    /// </exception>
+    /// <exception cref="DbUpdateConcurrencyException">
+    ///     Persisting data into database is not possibly due to concurrency issues.
     /// </exception>
     /// <returns>
     ///     Returns <see cref="Task" /> representing asynchronous operation which includes the result.
