+++ Utils/GuardsService.Exceptions/GuardsService.Exceptions.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <AssemblyName>EncyclopediaGalactica.Utils.GuardsService.Exceptions</AssemblyName>
+        <RootNamespace>EncyclopediaGalactica.Utils.GuardsService.Exceptions</RootNamespace>
+    </PropertyGroup>
+
+</Project>
