+++ Utils/GuardsService.Interfaces/GuardsService.Interfaces.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <AssemblyName>EncyclopediaGalactica.Utils.GuardsService.Interfaces</AssemblyName>
+        <RootNamespace>EncyclopediaGalactica.Utils.GuardsService.Interfaces</RootNamespace>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\GuardsService.Exceptions\GuardsService.Exceptions.csproj" />
+    </ItemGroup>
+
+</Project>
