+++ Utils/GuardsService/GuardsService.cs
@@ -1,36 +1,27 @@
 ﻿namespace EncyclopediaGalactica.Utils.GuardsService;
 
+using Exceptions;
+using Interfaces;
+
 public class GuardsService : IGuardsService
 {
     const string Msg = "Error happened while validating. For further information see inner exception.";
 
+    /// <inheritdoc />
     public void NotNull<T>(T val)
     {
-        try
-        {
-            if (val is not null) return;
-            throw new GuardsServiceValueShouldNoBeNullException($"The provided object value is null.");
-        }
-        catch (Exception e)
-        {
-            throw new GuardsServiceException(Msg, e);
-        }
+        if (val is not null) return;
+        throw new GuardsServiceValueShouldNoBeNullException($"The provided object value is null.");
     }
 
+    /// <inheritdoc />
     public void IsNotEqual(long providedValue, long comparedTo)
     {
-        try
+        if (providedValue == comparedTo)
         {
-            if (providedValue == comparedTo)
-            {
-                string msg;
-                msg = $"The provided value {providedValue} cannot be equal to: {comparedTo}";
-                throw new GuardsServiceValueShouldNotBeEqualToException(msg);
-            }
-        }
-        catch (Exception e)
-        {
-            throw new GuardsServiceException(Msg, e);
+            string msg;
+            msg = $"The provided value {providedValue} cannot be equal to: {comparedTo}";
+            throw new GuardsServiceValueShouldNotBeEqualToException(msg);
         }
     }
 }
