--- Utils/GuardsService/IGuardsService.cs
@@ -1,7 +1,0 @@
-namespace EncyclopediaGalactica.Utils.GuardsService;
-
-public interface IGuardsService
-{
-    void NotNull<T>(T val);
-    void IsNotEqual(long providedValue, long comparedTo);
-}
